pyproject.toml
[project]
name = "spectofind"
version = "0.1.0"
description = "Spectrogram-based environmental sound classifier using EfficientNet-B0 transfer learning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "torchaudio>=2.2.0",
    "timm>=0.9.16",
    "librosa>=0.10.1",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "pillow>=10.2.0",
    "requests>=2.31.0",
    "sounddevice>=0.4.6",
    "scikit-learn>=1.4.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.7",
]

[project.scripts]
spectofind-train   = "spectofind.train:main"
spectofind-infer   = "spectofind.infer:main"
spectofind-eval    = "spectofind.evaluate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/spectofind"]

[tool.uv]
dev-dependencies = []

[[tool.uv.index]]
name = "pytorch-cu121"
url  = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch        = { index = "pytorch-cu121" }
torchvision  = { index = "pytorch-cu121" }
torchaudio   = { index = "pytorch-cu121" }
